+++ test/Rakefile.rb
@@ -1,4 +1,0 @@
-
-task :shellcheck do
-  Rake.sh('shellcheck dummy_bash_script.sh')
-end