+++ test/integration/ide_work/good.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-abc="123"
+abc="123 456"
 echo "${abc}"